feat: fetchSongs
